move process_yearly_table to yearly_table.py
