+++ pipeline/main.py
@@ -26,17 +26,6 @@ logger.addHandler(file_handler)
 
 
 config = read_config("config.json")
-def process_yearly_table(product: ChelsaProduct,
-                         zonal_dir: Path,
-                         out_path: Path,
-                         sort_values: list[str]):
-    
-    yearly_table = yearly_table_generator(product=product,
-                                          zonal_dir=zonal_dir,
-                                          sort_values=sort_values)
-    
-    yearly_table.to_csv(out_path, encoding='utf-8', index=False)
-    
 
 # TODO: add overwrite that will replace the existing file if needed
